+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+[0.0.3] 2019-09-26
+
+-   Replace @emotion/styled with emotion jsx pragma and css prop
+
+[0.0.2] 2019-09-26
+
+-   Replace styled-system with @emotion/styled
+
 [0.0.1] 2019-07-31
 
 -   Initial pre-release